questdb/sender/requirements.txt
psycopg2-binary
numpy

